requirements.txt
Django==5.2.12
django-filter==25.2
djangorestframework==3.17.0
drf-spectacular==0.28.0
mysqlclient==2.2.8
django-cors-headers==4.9.0
gunicorn==25.1.0
